feat: extract shared pi host coordination helpers
- move attachment extraction and prompt inlining helpers into a shared harness-host module and keep only PI-specific prompt body assembly in pi.ts
- move runtime todo waiting-user coordination into a shared coordinator so question completion and blocked todo behavior no longer live directly in runPi()
- move workspace-boundary path and command validation into a shared helper module and keep PI wrappers focused on enforcement
- remove the old PI-side skill directory loading path and drop the now-unused workspace_skill_dirs request field from the PI host contract
- update harness-host, harness adapter, and ts-runner tests to validate workspace_skills-based plumbing and the extracted helper boundaries
- validation: node --import tsx --test src/contracts.test.ts src/pi.test.ts src/pi-runtime-tools.test.ts
- validation: node --import tsx --test src/harness-conformance.test.ts src/ts-runner.test.ts
- validation: node --import tsx --test ../harnesses/src/pi.test.ts
