+++ runtime/harnesses/src/pi.ts
@@ -44,7 +44,6 @@ export const piHarnessDefinition: HarnessDefinition = {
         timeout_seconds: params.timeoutSeconds,
         runtime_api_base_url: params.runtimeApiBaseUrl ?? null,
         system_prompt: params.runtimeConfig.system_prompt,
-        workspace_skill_dirs: params.workspaceSkills.map((skill) => skill.source_dir),
         workspace_skills: params.workspaceSkills.map((skill) => ({
           skill_id: skill.skill_id,
           skill_name: skill.skill_name,
